refactor: reorganize source
