+++ src/orcabridge/base.py
@@ -44,16 +44,16 @@ class Operation:
         return Invocation(self, invocation_id, streams)
 
     def __call__(self, *streams: "SyncStream") -> "SyncStream":
-        # if any source is passed in as a stream, invoke it to extract stream first
-        from .source import Source
+        # trigger call on source if passed as stream
 
-        streams = [s() if isinstance(s, Source) else s for s in streams]
-
+        streams = [
+            stream() if isinstance(stream, Source) else stream for stream in streams
+        ]
         output_stream = self.forward(*streams)
         # create an invocation instance
         invocation = self.get_invocation(*streams)
         # label the output_stream with the invocation information
-        output_stream.source = invocation
+        output_stream.invocation = invocation
 
         # delay import to avoid circular import
         from .tracker import Tracker
@@ -118,22 +118,22 @@ class Stream:
     A stream is a collection of tagged-packets that are generated by an operation.
     The stream is iterable and can be used to access the packets in the stream.
 
-    A stream has propery `source` that is an instance of Invocation that generated the stream.
+    A stream has propery `invocation` that is an instance of Invocation that generated the stream.
     This may be None if the stream is not generated by an operation.
     """
 
     def __init__(self):
-        self._source: Optional[Invocation] = None
+        self._invocation: Optional[Invocation] = None
 
     @property
-    def source(self) -> Optional[Invocation]:
-        return self._source
+    def invocation(self) -> Optional[Invocation]:
+        return self._invocation
 
-    @source.setter
-    def source(self, value: Invocation) -> None:
+    @invocation.setter
+    def invocation(self, value: Invocation) -> None:
         if not isinstance(value, Invocation):
-            raise TypeError("source must be an instance of Invocation")
-        self._source = value
+            raise TypeError("invocation field must be an instance of Invocation")
+        self._invocation = value
 
     def __iter__(self) -> Iterator[Tuple[Tag, Packet]]:
         raise NotImplementedError("Subclasses must implement __iter__ method")
@@ -146,9 +146,9 @@ class SyncStream(Stream):
     """
 
     def __hash__(self) -> int:
-        if self.source is not None:
+        if self.invocation is not None and hasattr(self.invocation, "invocation_id"):
             # use the invocation ID as the hash
-            return self.source.invocation_id
+            return self.invocation.invocation_id
         return super().__hash__()
 
     def keys(self) -> Tuple[List[str], List[str]]:
@@ -197,3 +197,32 @@ class SyncStream(Stream):
             return MapPackets(transformer)(self)
         elif isinstance(transformer, Callable):
             return transformer(self)
+
+    def __mul__(self, other: "SyncStream") -> "SyncStream":
+        """
+        Returns a new stream that is the result joining with the other stream
+        """
+        from .mapper import Join
+
+        if not isinstance(other, SyncStream):
+            raise TypeError("other must be a SyncStream")
+        return Join()(self, other)
+
+
+class Source(Operation, SyncStream):
+    """
+    A base class for all sources in the system. A source can be seen as a special
+    type of Operation that takes no input and produces a stream of packets.
+    For convenience, the source itself is also a stream and thus can be used
+    as an input to other operations directly.
+    """
+
+    def __init__(self) -> None:
+        super().__init__()
+        self._invocation = self
+
+    def __hash__(self):
+        return super(Operation, self).__hash__()
+
+    def __iter__(self) -> Iterator[Tuple[Tag, Packet]]:
+        yield from self()
